DROOLS-6448: improve tests
